After following all the instructions from FOSUserBundle and FOSFacebookBundle to set up an external facebook auth that triggers symfony auth, I ended up with the error :
There is no user provider for user "Acme\DemoBundle\Entity\User". 

when accessing /login_check (redirected from facebook.com)
The exception was thrown in the refreshUser method.

It seemed that at this point, $user had a username but no facebookId. Adding the proposed line solves the issue.
(but is it the good way to solve it ?...)